fix : reorganised files for easier future merges
